Fix job and config identification by matching with URL.
This builds up a list of jobs and activeConfigurations from the first
Jenkins API 'jobs' call and associates a root URL with each job and
config. Later on, tasks and jobs can match their URL against this list
to identify their job and configuration without any kind of funky
parsing.